ref(api): move static api functions to separate class
